requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
python-dotenv==0.17.1
